Fix obscure fluid pipe insertion behaviour
I noticed that the Fluid Rationer was being weird only in the context of
an ore processing setup. It would stop working whenever a new ore type
was processed, even when apparrently empty. I eventually determined that
since the output mode was set to NONE after ejection, the fluid pipes
could insert back into the output storage, causing an invalid state.

To fix this, I had to add a canInsert check in BlockPipeVertex and
actually use EXTRACT_ONLY, a mode that has existed unused since the
inception of the my fluid pipes eons ago.